shell::command_data

Enum Run

pub enum Run {
    Command(CommandWithArgs),
    BackgroundCommand(CommandWithArgs),
    Pipe(Vec<Run>),
    Sequence(Vec<Run>),
    And(Vec<Run>),
    Or(Vec<Run>),
    Subshell(Box<Run>),
    Empty,
}

Command(s) ready to run with context.

Variants

Command(CommandWithArgs)

BackgroundCommand(CommandWithArgs)

Pipe(Vec<Run>)

Sequence(Vec<Run>)

And(Vec<Run>)

Or(Vec<Run>)

Subshell(Box<Run>)

Empty

Implementations

impl Run

pub fn push_run(self, new_run: Run) -> Self

Push a new Run onto the Run. If it is not the first command will add to or create a sequence.

pub fn push_pipe(self, new_run: Run) -> Self

Push Run onto an existing or create a new pipe sequence.

pub fn push_sequence(self, new_run: Run) -> Self

Push new Run onto an existing or create a new sequence.

pub fn push_and(self, new_run: Run) -> Self

Push new Run onto an existing or create a new AND sequence.

pub fn push_or(self, new_run: Run) -> Self

Push new Run onto an existing or create a new OR sequence.

pub fn push_stdin_front(&mut self, fd: Option<FileDesc>)

If fd is Some value then put it at the front of the redir queue for the first command in the Run.

pub fn push_stdout_front(&mut self, fd: Option<FileDesc>)

If fd is Some value then put it at the front of the redir queue for the last command in the Run.

pub fn fds_to_internal(&mut self, fd_set: &HashSet<FileDesc>)

pub fn get_internal_fds(&self) -> HashSet<FileDesc>

Return a set of all the โinternalโ file descriptors (for pipes etc).

pub fn push_arg_end(&mut self, arg: Arg)

If fd is Some value then put it at the front of the redir queue for the last command in the Run.

pub fn extend_redirs_end(&mut self, redirs: &Redirects)

If fd is Some value then put it at the front of the redir queue for the last command in the Run.