Module shell::parse

Shell reader/parser. Parses a string into a shell command. Roughly a subset of https://www.gnu.org/software/bash/manual/html_node/Shell-Expansions.html

Structs

ParseState 🔒

ParsedJob

A list of commands to run in a pipe as a job.

Enums

SeqType 🔒

Type of the current sequence being parsed.

Functions

char_to_hex_num 🔒

consume_whitespace 🔒

escape_to_char 🔒

Read an ascii char from 0x00-0x7F endowed in a string as ‘\xXX’ where X is single hex digit.

parse_line

parse_line_inner 🔒

push_next_seq_item 🔒

push_next_seq_run 🔒

read_arg 🔒

read_chars_until 🔒

read_simple_string 🔒

Read string surrounded by single quote (’). Assumes chars is on the open quote and consumes the end quote. This simply reads the chars until the next ’ and puts them in a String Arg. Note, can not produce a string containing a ’ character.

read_special_arg 🔒

read_string 🔒

Read string surrounded by quote (“). Assumes chars is on the open quote and consumes the end quote. This version will handle interpolation and escape chars.

read_token 🔒

read_utf_scalar 🔒

Read a UTF8 codepoint encoded in a string with ‘\uXXXXXXXX’ or \u{XXXXXXXX}’ where X is a single hex digit. There can be 1-8 hex values (X- a nibble) in the encoding (Up to 4 bytes).