Module shell::platform::unix

Modules

umask 🔒

Macros

impl_is_minus_one 🔒

Structs

Sys

UnixFileDesc

Raw file descriptor for the target platform.

UnixPid

Process ID for the target platform.

UnixTermSettings

Saved terminal settings.

Constants

CLOEXEC_MSG_FOOTER 🔒

STDERR_FILENO

STDIN_FILENO

STDOUT_FILENO

Traits

IsMinusOne 🔒

Functions

arg_into_args 🔒

close_extra_fds 🔒 ⚠

cvt 🔒

exec 🔒

os2c 🔒

rlimit_to_c 🔒

send_error_to_parent 🔒 ⚠

Send an error code back tot he parent from a child process indicating it failed to fork. This function exits the child when done (i.e. never returns). ONLY call on error in a child process. SAFETY: this is doing “unsafe” libc/process stuff…

setup_group_term 🔒

Setup the process group for the current pid as well term if interactive. Call from both the parent and child proc to avoid race conditions.

Type Aliases

OsSignal

An OS signal.

RlimitResource

mode_t