shell::platform::unix

Function send_error_to_parent

unsafe fn send_error_to_parent(output: i32, err: Error)

Send an error code back tot he parent from a child process indicating it failed to fork. This function exits the child when done (i.e. never returns). ONLY call on error in a child process. SAFETY: this is doing “unsafe” libc/process stuff…