Module shell::platform::unix::umask

Structs

MaskType 🔒

Enums

PermissionOperator 🔒

Statics

NIX_PERMISSIONS 🔒

Functions

build_mask 🔒

decode_symbolic_mode_string 🔒

get_and_clear_umask

Cears the current umask and returns the previous umask.

get_class 🔒

get_perms 🔒

get_umask_tokens 🔒

is_digit 🔒

is_permission_token 🔒

is_user_access_token 🔒

make_parsable_octal_string 🔒

makes sure the returned string is 4 characters and the first character is 0.

merge_and_set_umask

If mask_string is a mode string then merge it with umask and set the current umask. If mask_string is an int then treat it as a umask and set the current umask (no merge)).

octal_string_to_mode 🔒

octal_string_to_mode_t 🔒

set_umask

Set current umask to umask.

to_mask_type 🔒

to_mode 🔒

with_umask_tokens 🔒

Type Aliases

mode_t