shell::platform::unix

Module umask

Structs

MaskType ğŸ”’

Enums

PermissionOperator ğŸ”’

Statics

NIX_PERMISSIONS ğŸ”’

Functions

build_mask ğŸ”’

decode_symbolic_mode_string ğŸ”’

get_and_clear_umask

Clears the current umask and returns the previous umask.

get_class ğŸ”’

get_perms ğŸ”’

get_umask_tokens ğŸ”’

is_digit ğŸ”’

is_permission_token ğŸ”’

is_user_access_token ğŸ”’

make_parsable_octal_string ğŸ”’

makes sure the returned string is 4 characters and the first character is 0.

merge_and_set_umask

If mask_string is a mode string then merge it with umask and set the current umask. If mask_string is an int then treat it as a umask and set the current umask (no merge)).

octal_string_to_mode ğŸ”’

octal_string_to_mode_t ğŸ”’

set_umask

Set current umask to umask.

to_mask_type ğŸ”’

to_mode ğŸ”’

with_umask_tokens ğŸ”’

Type Aliases

mode_t